If pretending do not reboot.
Former-commit-id: ea2bc6c00eccedbd68ae44e07ecbd78cf3a9fd8a [formerly 9be1af50fffaf18c6088c6c6f7857b447d51e2bc]
Former-commit-id: 16a0db74b99753e5e331e5509251e6454435ba7c

+++ minstall.cpp
@@ -2587,7 +2587,7 @@ void MInstall::gotoPage(int next)
         // finished
         updateCursor(Qt::WaitCursor);
         cleanup();
-        if (checkBoxExitReboot->isChecked()) {
+        if (!pretend && checkBoxExitReboot->isChecked()) {
             shell.run("/usr/local/bin/persist-config --shutdown --command reboot &");
         }
         qApp->exit(0);
